requirements.txt
chromadb==0.5.5
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
requests>=2.32.0
python-dotenv>=1.0.1
